requirements.txt
requests
google-genai
pandas
ta
python-dotenv
aiohttp
feedparser
solana
solders
base58
driftpy
anchorpy
langgraph
autogen-agentchat
qdrant-client
pydantic
